backend/rag-server/requirements.txt
python-dotenv
fastapi
uvicorn
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
